# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoTasks is a robust API for managing company tasks and collaborations.\nIt provides comprehensive features for:\n- **Company and Collaborator Management**: Onboard companies and manage staff with hierarchical roles.\n- **Activity and Task Lifecycle Tracking**: Create and monitor activities and tasks through various stages.\n- **Role-Based Access Control (RBAC)**: Secure endpoints with granular permissions for Admins, Ops, Owners, and Managers.\n- **Performance Ratings and Assignments**: Assign tasks to collaborators and rate their performance.\n- **Secure Authentication**: JWT-based authentication for secure access.",
        "title": "GoTasks API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/company/manager-owner/activity/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List activities in your company with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity | Company Manager or Owner"
                ],
                "summary": "List activities",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "title"
                        ],
                        "type": "string",
                        "default": "id",
                        "description": "Field to sort by",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "default": "ASC",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter expression",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_Paginated-internal_activity_CompanyActivityDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity | Company Manager or Owner"
                ],
                "summary": "Create activity",
                "parameters": [
                    {
                        "description": "Activity create payload",
                        "name": "activity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_activity.ActivityCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_activity_CompanyActivityDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/company/manager-owner/activity/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a single activity by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity | Company Manager or Owner"
                ],
                "summary": "Get activity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_activity_CompanyActivityDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity | Company Manager or Owner"
                ],
                "summary": "Update activity",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity update payload",
                        "name": "activity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_activity.ActivityUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_activity_CompanyActivityDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete activity by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity | Company Manager or Owner"
                ],
                "summary": "Delete activity",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/company/manager-owner/activity/{id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update status of an activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity | Company Manager or Owner"
                ],
                "summary": "Update activity status",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update payload",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_activity.ActivityUpdateStatusDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-tasks_internal_lifecycle.LifecycleStatusDTO": {
            "type": "string",
            "enum": [
                "CREATED",
                "IN_PROGRESS",
                "BLOCKED",
                "COMPLETED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "LifecycleStatusCreatedDTO",
                "LifecycleStatusInProgressDTO",
                "LifecycleStatusBlockedDTO",
                "LifecycleStatusCompletedDTO",
                "LifecycleStatusCancelledDTO"
            ]
        },
        "go-tasks_internal_pkg_httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "go-tasks_internal_pkg_httpx.MessageDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed"
                }
            }
        },
        "go-tasks_internal_pkg_httpx.Paginated-internal_activity_CompanyActivityDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_activity.CompanyActivityDTO"
                    }
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/go-tasks_internal_pkg_httpx.MessageDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_Paginated-internal_activity_CompanyActivityDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Paginated-internal_activity_CompanyActivityDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Response-internal_activity_CompanyActivityDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_activity.CompanyActivityDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_activity.ActivityAccountDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_activity.ActivityCollaboratorDTO": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/internal_activity.ActivityAccountDTO"
                },
                "account_id": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "internal_activity.ActivityCreateDTO": {
            "type": "object",
            "properties": {
                "lifecycle": {
                    "$ref": "#/definitions/internal_activity.ActivityLifecycleCreateDTO"
                },
                "owner_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_activity.ActivityLifecycleCreateDTO": {
            "type": "object",
            "properties": {
                "due_date": {
                    "type": "string"
                },
                "init_date": {
                    "type": "string"
                }
            }
        },
        "internal_activity.ActivityLifecycleDTO": {
            "type": "object",
            "properties": {
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "init_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_activity.ActivityStatusUpdateDTO"
                    }
                }
            }
        },
        "internal_activity.ActivityLifecycleUpdateDTO": {
            "type": "object",
            "properties": {
                "due_date": {
                    "type": "string"
                },
                "init_date": {
                    "type": "string"
                }
            }
        },
        "internal_activity.ActivityStatusUpdateDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "lifecycle_id": {
                    "type": "string"
                },
                "status_after": {
                    "type": "string"
                },
                "status_before": {
                    "type": "string"
                },
                "update_date": {
                    "type": "string"
                }
            }
        },
        "internal_activity.ActivityUpdateDTO": {
            "type": "object",
            "properties": {
                "lifecycle": {
                    "$ref": "#/definitions/internal_activity.ActivityLifecycleUpdateDTO"
                },
                "owner_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_activity.ActivityUpdateStatusDTO": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/go-tasks_internal_lifecycle.LifecycleStatusDTO"
                }
            }
        },
        "internal_activity.CompanyActivityDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "company_id": {
                    "type": "string"
                },
                "created_by": {
                    "$ref": "#/definitions/internal_activity.ActivityCollaboratorDTO"
                },
                "created_by_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lifecycle": {
                    "$ref": "#/definitions/internal_activity.ActivityLifecycleDTO"
                },
                "lifecycle_id": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/internal_activity.ActivityCollaboratorDTO"
                },
                "owner_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}